clean up dd logging
