+++ litellm/integrations/datadog/types.py
@@ -1,0 +1,6 @@
+class DatadogPayload(TypedDict, total=False):
+    ddsource: str
+    ddtags: str
+    hostname: str
+    message: str
+    service: str
